task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,22 +10,22 @@
  */
 int main(void)
 {
-	unsigned int prev, next, ans;
+	unsigned long long int prev, next, ans;
 	int i;
-	
+
 	i = 1;
 	prev = 1;
 	next = 2;
-	printf("%u, %u, ", prev, next);
+	printf("%llu, %llu, ", prev, next);
 	while (i <= 97)
 	{
 		ans = prev + next;
 		prev = next;
 		next = ans;
 		if (i == 97)
-			printf("%u\n", ans);
+			printf("%llu\n", ans);
 		else
-			printf("%u, ", ans);
+			printf("%llu, ", ans);
 		i++;
 	}
 	return (0);
